fix: token rates subscription

+++ apps/extension/src/core/domains/tokenRates/store.ts
@@ -1,22 +1,93 @@
-import { DEBUG } from "@core/constants"
 import { db } from "@core/db"
+import { unsubscribe } from "@core/handlers/subscriptions"
+import { log } from "@core/log"
 import { chaindataProvider } from "@core/rpcs/chaindata"
+import { Port } from "@core/types/base"
+import { TokenList } from "@talismn/chaindata-provider"
 import { fetchTokenRates } from "@talismn/token-rates"
+import { Subscription, liveQuery } from "dexie"
+import { debounce } from "lodash"
+import { BehaviorSubject, Subject } from "rxjs"
 
-const minimumHydrationInterval = 60_000 // 60_000ms = 60s = 1 minute
+const MIN_REFRESH_INTERVAL = 20_000 // 60_000ms = 60s = 1 minute
+const REFRESH_INTERVAL = 60_000 // 5 minutes
 
 export class TokenRatesStore {
-  #lastHydratedAt = 0
+  #lastUpdateTokenIds = ""
+  #lastUpdateAt = Date.now() // will prevent a first empty call if tokens aren't loaded yet
+  #subscriptions = new BehaviorSubject<string[]>([])
+  #isWatching = false
+
+  constructor() {
+    this.watchSubscriptions()
+  }
+
+  /**
+   * Toggles on & off the price updates, based on if there are any active subscriptions
+   */
+  private watchSubscriptions = (): void => {
+    let pollInterval: NodeJS.Timer | null = null
+    let subTokenList: Subscription | null = null
+
+    this.#subscriptions.subscribe((subscriptions: string[]) => {
+      if (subscriptions.length) {
+        // watching state check
+        if (this.#isWatching) return
+        this.#isWatching = true
+
+        // refresh price every minute if observed
+        pollInterval = setInterval(() => {
+          if (this.#subscriptions.observed) this.hydrateStore()
+        }, REFRESH_INTERVAL)
+
+        // refresh when token list changes : crucial for first popup load after install or db migration
+        const obsTokens = liveQuery(() => chaindataProvider.tokens())
+        subTokenList = obsTokens.subscribe(
+          debounce(async (tokens) => {
+            if (this.#subscriptions.observed) await this.updateTokenRates(tokens)
+          }, 500) // debounce to delay in case on first load first token list is empty
+        )
+      } else {
+        // watching state check
+        if (!this.#isWatching) return
+        this.#isWatching = false
+
+        if (pollInterval) {
+          clearInterval(pollInterval)
+          pollInterval = null
+        }
+
+        if (subTokenList) {
+          subTokenList.unsubscribe()
+          subTokenList = null
+        }
+      }
+    })
+  }
 
   async hydrateStore(): Promise<boolean> {
-    const now = Date.now()
-    if (now - this.#lastHydratedAt < minimumHydrationInterval) return false
-
     try {
-      // update tokenRates for known tokens
       const tokens = await chaindataProvider.tokens()
-      const tokenRates = await fetchTokenRates(tokens)
-      db.tokenRates.bulkPut(
+      await this.updateTokenRates(tokens)
+
+      return true
+    } catch (error) {
+      log.error(`Failed to fetch tokenRates`, error)
+      return false
+    }
+  }
+
+  private async updateTokenRates(tokens: TokenList): Promise<void> {
+    const now = Date.now()
+    const strTokenIds = Object.keys(tokens ?? {}).join(",")
+    if (now - this.#lastUpdateAt < MIN_REFRESH_INTERVAL && this.#lastUpdateTokenIds === strTokenIds)
+      return
+
+    const tokenRates = await fetchTokenRates(tokens)
+
+    await db.transaction("rw", db.tokenRates, async (tx) => {
+      // override all tokenRates
+      await db.tokenRates.bulkPut(
         Object.entries(tokenRates).map(([tokenId, tokenRates]) => ({
           tokenId,
           rates: tokenRates,
@@ -25,17 +96,25 @@ export class TokenRatesStore {
 
       // delete tokenRates for tokens which no longer exist
       const tokenIds = await db.tokenRates.toCollection().primaryKeys()
-      await db.tokenRates.bulkDelete(tokenIds.filter((tokenId) => tokens[tokenId] === undefined))
+      if (tokenIds.length)
+        await db.tokenRates.bulkDelete(tokenIds.filter((tokenId) => tokens[tokenId] === undefined))
+    })
 
-      // update lastHydratedAt
-      this.#lastHydratedAt = now
+    // update lastHydratedAt
+    this.#lastUpdateAt = now
+    this.#lastUpdateTokenIds === strTokenIds
+  }
 
-      return true
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      DEBUG && console.error(`Failed to fetch tokenRates`, error)
-      return false
-    }
+  public subscribe(id: string, port: Port): void {
+    this.#subscriptions.next([...this.#subscriptions.value, id])
+
+    this.hydrateStore()
+
+    // close subscription
+    port.onDisconnect.addListener((): void => {
+      unsubscribe(id)
+      this.#subscriptions.next(this.#subscriptions.value.filter((subId) => subId !== id))
+    })
   }
 }
